requirements.txt
fastapi
uvicorn[standard]
rembg
pillow
numpy
